requirements.txt
uvicorn 
fastapi
fastapi[standard]
sqlalchemy
pyodbc
driver ODBC
fastapi-cache2